Depth-First ProNoC Implementation with OpenPiton Support
This patch brings support of the (3.5D) "Multi-Mesh" topologies with the
Depth-First routing algorithm in ProNoC. Compared to its original formulation,
the algorithm has been slightly modified to support a single source of credits
for packets injection / reception. The routing algorithm can be evaluated either
within the ProNoC infrastructure, or with OpenPiton, using the modules in
src_multi_mesh/wrapper.sv as the interface between the ProNoC routers and the
OpenPiton endpoints (tiles). For OpenPiton, the physical NoCs generator now
creates 3 (Phy-)NoCs for each OpenPiton chip.